requirements.txt
google-api-python-client==2.186.0
google-auth-httplib2==0.2.1
google-auth-oauthlib==1.2.3
langgraph==0.2.59
langgraph-checkpoint-postgres==2.0.8
langchain-ollama==0.2.2
langchain-core==0.3.28
slack-sdk==3.33.4
opentelemetry-api==1.22.0
opentelemetry-sdk==1.22.0
opentelemetry-exporter-otlp==1.22.0
opentelemetry-instrumentation==0.43b0
psycopg[binary]>=3.2.0,<4.0.0
pyyaml==6.0.1

# MCP (Model Context Protocol)
mcp>=1.0.0

# Testing
pytest==8.3.4
pytest-cov==6.0.0

